requirements.txt
psycopg[binary]>=3.1.18
fastmcp>=1.0.0,<2.0.0
pydantic>=2.6.0

fastapi
uvicorn[standard]
sqlmodel #not installed yet


